feat: add useMemo
