+++ src/client/userProfile/_user.variables.scss
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
+@import '../../shared/styles/scss/_shared';
 
 $FONT: 'Roboto', sans-serif;
