+++ src/client/userProfile/components/Balance/Balance.module.scss
@@ -1,4 +1,4 @@
-@import '../user.variables';
+@import '../../_user.variables';
 
 .balanceContainer {
   box-sizing: border-box;
