+++ src/client/userProfile/components/SteamConnection/SteamConnection.module.scss
@@ -1,4 +1,4 @@
-@import '../user.variables';
+@import '../../_user.variables';
 
 .steamContainer {
   width: 270px;
