+++ src/client/userProfile/components/TeamUserProfile/TeamUserProfil.module.scss
@@ -1,4 +1,4 @@
-@import '../user.variables';
+@import '../../_user.variables';
 
 .team {
   width: 270px;
@@ -55,6 +55,7 @@
 
 .name {
   font-size: 14px;
+    margin-bottom: 10px;
 }
 
 .date {
@@ -67,9 +68,10 @@
   background: none;
   border: none;
   outline: none;
-  margin-top: 17px;
+  margin-top: 22px;
   padding: 0;
   font-size: 12px;
+  cursor: pointer;
 }
 
 .leaveTeam:focus {
