+++ src/client/userProfile/components/TournamentHistory/TournamentHistory.jsx
@@ -1,15 +1,15 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 
 import styles from './TournamentHistory.module.scss';
-import Button from '../../../shared/components/Button';
+import Button from '../../../../shared/components/Button';
 
 const TournamentHistory = props => {
   const { tournaments } = props;
 
   const tournamentElements = tournaments.map(
-    ({ name, date, mode, role, result }) => (
-      <tr className={styles.tournamentItem}>
+    ({ name, date, mode, role, result }, number) => (
+      <tr className={styles.tournamentItem} key={`${name}${date}${number}`}>
         <td>{name}</td>
         <td>{date}</td>
         <td>{mode}</td>
@@ -18,10 +18,11 @@ const TournamentHistory = props => {
       </tr>
     ),
   );
-  return (
+  return useMemo(
+    () => (
       <div className={styles.container}>
-        <div className={styles.tournaments}>
-          <table className={styles.container}>
+        <table className={styles.tournaments}>
+          <tbody className={styles.container}>
             <tr className={styles.title}>
               <th>TOURNAMENT</th>
               <th>DATE</th>
@@ -30,12 +31,19 @@ const TournamentHistory = props => {
               <th>RESULT</th>
             </tr>
             {tournamentElements}
-          </table>
+          </tbody>
+        </table>
+        <div className={styles.buttonContainer}>
+          <Button
+            classType="outline"
+            text="Load More"
+            additionalClass={styles.loadMore}
+            handleClick={() => console.log('LOAD MORE')}
+          />
         </div>
-          <div className={styles.buttonContainer}>
-              <Button classType='outline' text="Load More" additionalClass={styles.loadMore}/>
-          </div>
       </div>
+    ),
+    [tournamentElements],
   );
 };
