+++ src/client/userProfile/components/UserProfile/UserProfile.jsx
@@ -8,16 +8,9 @@ import PlayerOverview from '../PlayerOverview';
 import SteamConnection from '../SteamConnection';
 import Balance from '../Balance';
 
-const teamProps = {
-  name: 'JabbasTeam 2',
-  date: 'Nov 3, 2020 13:47',
-  charge: 5,
-  invites: 6,
-};
-
 const UserProfile = props => {
-  const { player, balance, tournaments } = props;
-  const { playerName, playerItems, editProfileButton } = player;
+  const { player, balance, tournaments, team } = props;
+  const { playerName, playerItems } = player;
   const { charge, money } = balance;
 
   return (
@@ -25,11 +18,7 @@ const UserProfile = props => {
       <h2 className={styles.title}>Profile</h2>
 
       <p className={styles.subTitle}>Overview</p>
-      <PlayerOverview
-        playerName={playerName}
-        playerItems={playerItems}
-        editProfileButton={editProfileButton}
-      />
+      <PlayerOverview playerName={playerName} playerItems={playerItems} />
 
       <p className={styles.subTitle}>Steam connection</p>
       <SteamConnection />
@@ -38,13 +27,7 @@ const UserProfile = props => {
       <Balance charge={charge} money={money} />
 
       <p className={styles.subTitle}>Team</p>
-      <TeamUserProfile
-        name={teamProps.name}
-        date={teamProps.date}
-        charge={teamProps.charge}
-        invites={teamProps.invites}
-        team={teamProps.team}
-      />
+      <TeamUserProfile team teamProps={team} />
 
       <p className={styles.subTitle}>Tournament History</p>
       <TournamentHistory tournaments={tournaments} />
@@ -64,10 +47,13 @@ UserProfile.propTypes = {
 
   balance: PropTypes.objectOf(PropTypes.string),
   tournaments: PropTypes.arrayOf(PropTypes.objectOf(PropTypes.string)),
+
+  team: PropTypes.objectOf(PropTypes.string),
 };
 
 UserProfile.defaultProps = {
   player: {},
   balance: {},
   tournaments: [],
+  team: {},
 };
