+++ src/client/userProfile/components/UserProfile/UserProfile.module.scss
@@ -1,4 +1,4 @@
-@import '../user.variables';
+@import '../../_user.variables';
 
 p {
   margin: 0;
